move extractor factory -> path_matcher_factory
